refactor(indexer): extract BME helpers and add block results cache fallback
Move BME parsing helpers and types to dedicated files. Replace hard
throw on missing block results with a fallback fetch from the node
to prevent crashes when the cache was built before BME support.
